django_backend/requirements.txt
amqp==1.4.9
ansible==4.10.0
ansible-core==2.11.12
anyjson==0.3.3
bcrypt==4.0.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
concurrent-log-handler==0.9.16
ConcurrentLogHandler==0.9.1
cryptography==40.0.2
decorator==5.1.1
Django==2.1.1
django-cas-ng==4.1.1
django-celery==3.3.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gmssl==3.2.2
id-validator==1.0.20
idna==3.10
Jinja2==3.0.3
kombu==3.0.37
lxml==5.3.0
MarkupSafe==2.0.1
multi-key-dict==2.0.3
packaging==21.3
paramiko==3.5.0
pbr==6.1.0
phone==0.4.3
portalocker==2.7.0
pycparser==2.21
pycryptodome==3.21.0
pycryptodomex==3.21.0
PyJWT==1.7.1
PyMySQL==0.9.2
PyNaCl==1.5.0
pyparsing==3.1.4
python-cas==1.6.0
python-dateutil==2.9.0.post0
python-jenkins==1.8.2
python-stdnum==1.20
pytz==2024.2
PyYAML==6.0.1
pyzabbix==1.3.1
redis==2.10.6
requests==2.27.1
resolvelib==0.5.4
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.1.1
urllib3==1.26.20
validator.py==1.3.0
validators==0.20.0
xmltodict==0.14.2
